vagrant: add heartbeats / retries
When internet connection on user host (e.g. laptop) goes off and
a vagrant.up operation was in progress, eventual result of the
operation may be lost  due to kite ignoring send errors.

For now adding heartbeats for terraformplugin/vagrant <-> klient/vagrant
communication as it is a blanket cover for all networking edge-cases
that are not currently handled with kite.

The proper fix would be in the longer term to make kite guarantee
any successfful call is always sent (e.g. by queueing messages
during outages).